# contracts/openapi/ipv-gc.yaml
openapi: 3.1.0
info:
  title: IPV Gestión de Costos API
  version: 0.9.0
  description: |
    Contrato Fases 3–9. Autenticación Bearer JWT (EdDSA).
    Errores `application/problem+json`. Mutaciones aceptan `Idempotency-Key` y `If-Match`.
    Sync: POST /sync/push, GET /sync/changes, GET /sync/bootstrap.
    Licencias: Fake LicenseProvider (KEYGEN_MODE=FAKE). Cloud bloqueado hasta D-05.
    Tasas: ExchangeRateProvider + MOCK/SEED; API live bloqueada hasta D-04.
servers:
  - url: /api/v1
...
paths:
  /auth/login:
    post:
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                organization_id: { type: string, format: uuid }
      responses:
        "200":
          description: Tokens o desafío MFA
        "401":
          description: Credenciales inválidas (respuesta uniforme)
        "429":
          description: Throttle de login (IP o correo)
...
  /cost-sheets/{sheetId}/versions/{versionId}:
    get:
      tags: [cost-sheets]
      parameters:
        - in: path
          name: sheetId
          required: true
          schema: { type: string, format: uuid }
        - in: path
          name: versionId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Versión (ETag = version)
          headers:
            ETag:
              schema: { type: string }
    patch:
      tags: [cost-sheets]
      parameters:
        - in: header
          name: If-Match
          schema: { type: string }
      responses:
        "200": { description: Borrador actualizado }
        "409": { description: Congelado o conflicto de versión }
...
  /cost-sheets/{sheetId}/versions/{versionId}/submit:
    post:
      tags: [cost-sheets]
      responses:
        "200": { description: EN_REVISION, instantánea de tasas congelada }
        "422": { description: Reglas bloqueantes }
  /cost-sheets/{sheetId}/versions/{versionId}/validate:
    post:
      tags: [cost-sheets]
      responses:
        "200": { description: VALIDADA }
        "403": { description: Cuatro ojos }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []